+++ events-analyser/src/engine/elements/algorithm.rs
@@ -4,16 +4,20 @@ use std::{fmt::Debug, ops::Deref};
 
 use crate::engine::{
     Array, FlatWaveform, FlattenableWithIndex,
-    utils::HasName,
+    HasName,
     values::{Value, ValueError},
 };
 
-///
+/// Encapsulates all properties which defines a `trace-to-events` algorithm,
+/// used in event detection. The evaluator presumes these settings were used to
+/// detect the events it is evaluating in a particular bucket.
 #[derive(Debug, Deserialize, Clone)]
 #[serde(rename_all = "kebab-case")]
 pub(crate) struct Algorithm {
+    // Name of the algorithm properties.
     pub(crate) name: String,
     #[serde(flatten)]
+    // Properties defining the algorithm.
     pub(crate) properties: AlgorithmProperties,
 }
 
@@ -31,17 +35,19 @@ impl Deref for Algorithm {
     }
 }
 
-///
+/// Encapsulates all properties which defines a `trace-to-events` algorithm,
+/// used in event detection. The evaluator presumes these settings were used to
+/// detect the events it is evaluating in a particular bucket.
 #[derive(Debug, Deserialize, Clone)]
 #[serde(rename_all = "kebab-case")]
 pub(crate) enum AlgorithmProperties {
-    ///
+    /// Specifies the fixed threshold discriminator algorithm.
     FixedThreshold {
-        ///
+        /// The threshold property used in the event detection.
         threshold: Value<f64>,
-        ///
+        /// The duration property used in the event detection.
         duration: Value<Time>,
-        ///
+        /// The cool_down property used in the event detection.
         cool_down: Value<Time>,
     },
 }
@@ -71,21 +77,32 @@ impl FlattenableWithIndex for AlgorithmProperties {
     }
 }
 
+/// The version of `AlgorithmProperties` with all dependencies resolved.
 #[derive(Debug, Clone)]
 pub(crate) enum FlatAlgorithm {
+    /// Specifies the fixed threshold discriminator algorithm.
     FixedThreshold {
+        /// The threshold property used in the event detection.
         threshold: f64,
+        /// The duration property used in the event detection.
         duration: Time,
+        /// The cool_down property used in the event detection.
         cool_down: Time,
     },
 }
 
 impl FlatAlgorithm {
+    /// Estimates whether the given (time,intensity) pair could been detected by this algorithm
+    /// applied to a given [FlatWaveform] with given peak time and peak intensity.
+    /// # Parameters
+    /// - waveform: Waveform to model the pulse by.
+    /// - detected: Waveform to model the pulse by.
     pub(crate) fn is_true_positive(
         &self,
         waveform: &FlatWaveform,
-        intensity: Intensity,
-        dist: u32,
+        detected: (Time, Intensity),
+        pulse_peak: (Time, Intensity),
+        //dist: u32,
     ) -> bool {
         match self {
             &FlatAlgorithm::FixedThreshold {
@@ -93,13 +110,13 @@ impl FlatAlgorithm {
                 duration,
                 cool_down,
             } => {
-                let _height = threshold / intensity as f64;
+                let _height = threshold / pulse_peak.1 as f64;
                 let width = match waveform {
                     &FlatWaveform::Flat { width } => width,
                     &FlatWaveform::Triangular { base_width } => base_width,
                     &FlatWaveform::Gaussian { sd } => sd,
                 };
-                (dist as f64) < (duration + cool_down) as f64 + width
+                (detected.0 as f64 - pulse_peak.0 as f64).abs() < (duration + cool_down) as f64 + width
             }
         }
     }
